requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
click==8.2.1
    # via typer
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
pygments==2.19.2
    # via rich
pyyaml==6.0.2
    # via aws-login (setup.py)
rich==14.1.0
    # via
    #   aws-login (setup.py)
    #   typer
shellingham==1.5.4
    # via typer
typer==0.16.0
    # via aws-login (setup.py)
typing-extensions==4.14.1
    # via typer
